+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/cheatcodes/l1_handler_execute.rs
@@ -1,5 +1,5 @@
 use crate::{
-    runtime_extensions::call_to_blockifier_runtime_extension::rpc::{CallResult, call_l1_handler},
+    runtime_extensions::outer_call_runtime_extension::rpc::{CallResult, call_l1_handler},
     state::CheatnetState,
 };
 use blockifier::execution::syscalls::hint_processor::SyscallHintProcessor;
