+++ crates/cheatnet/src/runtime_extensions/mod.rs
@@ -1,4 +1,3 @@
-pub mod call_to_blockifier_runtime_extension;
 pub mod cheatable_starknet_runtime_extension;
 pub mod common;
 pub mod deprecated_cheatable_starknet_extension;
@@ -6,3 +5,4 @@ pub mod forge_config_extension;
 pub mod forge_runtime_extension;
 #[cfg(feature = "cairo-native")]
 mod native;
+pub mod outer_call_runtime_extension;
