+++ crates/cheatnet/src/runtime_extensions/outer_call_runtime_extension/execution/syscall_hooks.rs
