+++ crates/cheatnet/src/runtime_extensions/outer_call_runtime_extension/panic_parser.rs
