+++ crates/cheatnet/tests/cheatcodes/test_environment.rs
@@ -2,10 +2,10 @@ use crate::common::assertions::ClassHashAssert;
 use crate::common::{call_contract, deploy, library_call_contract};
 use crate::common::{deploy_contract, state::create_cached_state};
 use blockifier::state::cached_state::CachedState;
-use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::CallResult;
 use cheatnet::runtime_extensions::forge_runtime_extension::cheatcodes::declare::declare;
 use cheatnet::runtime_extensions::forge_runtime_extension::cheatcodes::storage::selector_from_name;
 use cheatnet::runtime_extensions::forge_runtime_extension::contracts_data::ContractsData;
+use cheatnet::runtime_extensions::outer_call_runtime_extension::rpc::CallResult;
 use cheatnet::state::{CheatnetState, ExtendedStateReader};
 use starknet_api::core::ClassHash;
 use starknet_api::core::ContractAddress;
